backend/requirements.txt
django
django-cors-headers
djangorestframework
drf-spectacular
gunicorn